+++ client/src/routes/Meetups/Meetups.module.scss
@@ -23,7 +23,7 @@
   }
 }
 
-.filter-sort {
+.filter-order {
   display: flex;
   justify-content: space-between;
   margin-bottom: 2rem;
